Refactor: move remaining posthog.client code
