+++ posthog/clickhouse/client/__init__.py
@@ -1,3 +1,9 @@
+from posthog.clickhouse.client.execute import query_with_columns, substitute_params, sync_execute
 from posthog.clickhouse.client.execute_async import execute_with_progress
 
-__all__ = ["execute_with_progress"]
+__all__ = [
+    "sync_execute",
+    "query_with_columns",
+    "substitute_params",
+    "execute_with_progress",
+]
