+++ posthog/client.py
@@ -1,220 +1,2 @@
-import json
-import types
-from functools import lru_cache
-from time import perf_counter
-from typing import Any, Dict, List, Optional, Sequence, Union
-
-import sqlparse
-from clickhouse_driver import Client as SyncClient
-from clickhouse_driver.util.escape import escape_params
-from django.conf import settings as app_settings
-from statshog.defaults.django import statsd
-
-from posthog.clickhouse.client.connection import ch_pool
-from posthog.clickhouse.query_tagging import get_query_tag_value, get_query_tags
-from posthog.errors import wrap_query_error
-from posthog.settings import TEST
-from posthog.utils import generate_short_id, patchable
-
-InsertParams = Union[list, tuple, types.GeneratorType]
-NonInsertParams = Dict[str, Any]
-QueryArgs = Optional[Union[InsertParams, NonInsertParams]]
-
-
-@lru_cache(maxsize=1)
-def default_settings() -> Dict:
-    from posthog.version_requirement import ServiceVersionRequirement
-
-    # On CH 22.3 we need to disable optimize_move_to_prewhere due to a bug. This is verified fixed on 22.8 (LTS),
-    # so we only disable on versions below that.
-    # This is calculated once per deploy
-    clickhouse_at_least_228, _ = ServiceVersionRequirement(
-        service="clickhouse", supported_version=">=22.8.0"
-    ).is_service_in_accepted_version()
-    if clickhouse_at_least_228:
-        return {}
-    else:
-        return {"optimize_move_to_prewhere": 0}
-
-
-def validated_client_query_id() -> Optional[str]:
-    client_query_id = get_query_tag_value("client_query_id")
-    client_query_team_id = get_query_tag_value("team_id")
-
-    if client_query_id and not client_query_team_id:
-        raise Exception("Query needs to have a team_id arg if you've passed client_query_id")
-    random_id = generate_short_id()
-    return f"{client_query_team_id}_{client_query_id}_{random_id}"
-
-
-@patchable
-def sync_execute(
-    query,
-    args=None,
-    settings=None,
-    with_column_types=False,
-    flush=True,
-):
-    if TEST and flush:
-        try:
-            from posthog.test.base import flush_persons_and_events
-
-            flush_persons_and_events()
-        except ModuleNotFoundError:  # when we run plugin server tests it tries to run above, ignore
-            pass
-
-    with ch_pool.get_client() as client:
-        start_time = perf_counter()
-
-        prepared_sql, prepared_args, tags = _prepare_query(client=client, query=query, args=args)
-
-        settings = {**default_settings(), **(settings or {}), "log_comment": json.dumps(tags, separators=(",", ":"))}
-
-        try:
-            result = client.execute(
-                prepared_sql,
-                params=prepared_args,
-                settings=settings,
-                with_column_types=with_column_types,
-                query_id=validated_client_query_id(),
-            )
-        except Exception as err:
-            err = wrap_query_error(err)
-            statsd.incr("clickhouse_sync_execution_failure", tags={"failed": True, "reason": type(err).__name__})
-
-            raise err
-        finally:
-            execution_time = perf_counter() - start_time
-
-            statsd.timing("clickhouse_sync_execution_time", execution_time * 1000.0)
-
-            if app_settings.SHELL_PLUS_PRINT_SQL:
-                print("Execution time: %.6fs" % (execution_time,))
-    return result
-
-
-def query_with_columns(
-    query: str,
-    args: Optional[QueryArgs] = None,
-    columns_to_remove: Optional[Sequence[str]] = None,
-    columns_to_rename: Optional[Dict[str, str]] = None,
-) -> List[Dict]:
-    if columns_to_remove is None:
-        columns_to_remove = []
-    if columns_to_rename is None:
-        columns_to_rename = {}
-    metrics, types = sync_execute(query, args, with_column_types=True)
-    type_names = [key for key, _type in types]
-
-    rows = []
-    for row in metrics:
-        result = {}
-        for type_name, value in zip(type_names, row):
-            if isinstance(value, list):
-                value = ", ".join(map(str, value))
-            if type_name not in columns_to_remove:
-                result[columns_to_rename.get(type_name, type_name)] = value
-
-        rows.append(result)
-
-    return rows
-
-
-def substitute_params(query, params):
-    """
-    Helper method to ease rendering of sql clickhouse queries progressively.
-    For example, there are many places where we construct queries to be used
-    as subqueries of others. Each time we generate a subquery we also pass
-    up the "bound" parameters to be used to render the query, which
-    otherwise only happens at the point of calling clickhouse via the
-    clickhouse_driver `Client`.
-
-    This results in sometimes large lists of parameters with no relevance to
-    the containing query being passed up. Rather than do this, we can
-    instead "render" the subqueries prior to using as a subquery, so our
-    containing code is only responsible for it's parameters, and we can
-    avoid any potential param collisions.
-    """
-    escaped = escape_params(params)
-    return query % escaped
-
-
-@patchable
-def _prepare_query(client: SyncClient, query: str, args: QueryArgs):
-    """
-    Given a string query with placeholders we do one of two things:
-
-        1. for a insert query we just format, and remove comments
-        2. for non-insert queries, we return the sql with placeholders
-        evaluated with the contents of `args`
-
-    We also return `tags` which contains some detail around the context
-    within which the query was executed e.g. the django view name
-
-    NOTE: `client.execute` would normally handle substitution, but
-    because we want to strip the comments to make it easier to copy
-    and past queries from the `system.query_log` easily with metabase
-    (metabase doesn't show new lines, so with comments, you can't get
-    a working query without exporting to csv or similar), we need to
-    do it manually.
-
-    We only want to try to substitue for SELECT queries, which
-    clickhouse_driver at this moment in time decides based on the
-    below predicate.
-    """
-    prepared_args: Any = QueryArgs
-    if isinstance(args, (list, tuple, types.GeneratorType)):
-        # If we get one of these it means we have an insert, let the clickhouse
-        # client handle substitution here.
-        rendered_sql = query
-        prepared_args = args
-    elif not args:
-        # If `args` is not truthy then make prepared_args `None`, which the
-        # clickhouse client uses to signal no substitution is desired. Expected
-        # args balue are `None` or `{}` for instance
-        rendered_sql = query
-        prepared_args = None
-    else:
-        # Else perform the substitution so we can perform operations on the raw
-        # non-templated SQL
-        rendered_sql = client.substitute_params(query, args)
-        prepared_args = None
-
-    formatted_sql = sqlparse.format(rendered_sql, strip_comments=True)
-    annotated_sql, tags = _annotate_tagged_query(formatted_sql, args)
-
-    if app_settings.SHELL_PLUS_PRINT_SQL:
-        print()
-        print(format_sql(formatted_sql))
-
-    return annotated_sql, prepared_args, tags
-
-
-def _annotate_tagged_query(query, args):
-    """
-    Adds in a /* */ so we can look in clickhouses `system.query_log`
-    to easily marry up to the generating code.
-    """
-    tags = get_query_tags()
-    # Annotate the query with information on the request/task
-    if "kind" in tags:
-        user_id = f" user_id:{tags['user_id']}" if "user_id" in tags else ""
-        query = f"/*{user_id} {tags.get('kind')}:{tags.get('id', '').replace('/', '_')} */ {query}"
-
-    return query, tags
-
-
-def format_sql(rendered_sql, colorize=True):
-    formatted_sql = sqlparse.format(rendered_sql, reindent_aligned=True)
-    if colorize:
-        try:
-            import pygments.formatters
-            import pygments.lexers
-
-            return pygments.highlight(
-                formatted_sql, pygments.lexers.get_lexer_by_name("sql"), pygments.formatters.TerminalFormatter()
-            )
-        except:
-            pass
-
-    return formatted_sql
+# File kept around for legacy purposes
+from posthog.clickhouse.client import *  # noqa T401
